Print "execution finished" banner from run.sh
Move the "execution finished" banner from the client JS to run.sh (in
the Docker container). This way, all the banners are printed in the same
place, and we have better control over when the banner is printed. This
prints the banner immediately after the code finishes running, then
pauses for 0.1 seconds, so that we can get the output of any
unsynchronized child processes to show up, even if they print after the
parent exits.