Fix num processes limit, reduce num allowed fds
Previously, I was using ulimit to restrict the number of processes in
each container, but, as it turns out, ulimit applies this restriciton
based on the uid, and because every container runs code as uid 999,
there are only 24 processes allowed *across all containers*. docker run
provides a --pids-limit flag that solves this problem.

This also reduces the number of processes to 16 (we aren't providing
many cores, so users shouldn't have many processes anyways), and the
number of file descriptors per process to 64. This means there can be at
most 1024 fds per container, to avoid ddos via system fd exhaustion.