Set 32MB filesystem quota, disable networking
This was a hell of a goose chase. Disabling networking was easy; setting
a filesystem quota was *hard*.

This sets the container filesystem to be read-only, mounts a separate
tmpfs (ramdisk) at the home directory /cppfiddle, and then bind-mounts
the source C/C++ file as a single read-only file within that directory.
The --tmpfs flag was much easier to work with than the --mount flag, as
the former allows passing arbitrary flags to the tmpfs engine, whereas
the --mount flag currently has no way to allow executable files in the
tmpfs mount. (This is not documented *anywhere*. -_-)