Add embeddable functionality
This adds a path /embed that can be put in an iframe, as well as changes
the styles so that the embedded view is friendlier to smaller window
sizes. There is a code-only view, a terminal-only view, and the split
view is kept in case anyone wants it (but isn't the default in the
embedded page).