Add support for including external files
This is a first attempt at allowing users to upload zip files containing
code that can be used by the program (either to link in extra files, or
to add files to the container that the program can use as input). It's
pretty janky but will be improved in the future.