Implement container timeouts in node, add CPU limit
Instead of doing timeouts within the container (using `timeout`), this
commit has node monitor everything from outside the container. This
makes timeouts a little more robust (e.g. in a forkbomb situation,
sometimes the timeout process was getting OOM-killed and then it didn't
effectively kill any processes), and it also allows us to do fancier
things with timeouts. In this commit, we extend the timeout every time
there is user interaction. (This is useful if you're testing a shell or
game or interactive program and want to be able to use a low-intensity
program for an extended amount of time.)

This commit also adds a container-wide CPU time limit (as opposed to a
per-process CPU time limit). If the sum of all processes in the
container use more than 15 seconds of CPU time, the container will be
killed. (This helps mitigate forkbombs and such.)