Implement Terminal, make programs runnable
This commit adds a bunch of code to implement the Terminal component
using xterm.js. This component serves as a React wrapper for the
functionality we are looking from xterm. It leverages a lot of new code
written in server-comm.ts to bind to a websocket, sending data that is
typed into the terminal and displaying received data.