Resize Ace editor after showing/hiding settings sidebar
Since showing/hiding the settings sidebar affects the width of the
editor div, and since Ace isn't smart enough to detect changes in the
size of its parent, we need to explicitly tell it to resize.

This commit adds a ref to Editor so that we can get its AceEditor
child component, and, from that, access the underlying `editor` object.
We can call the standard `resize()` method on that.