Make /proc/cplayground readable by unprivileged user
We don't want the cplayground user (i.e. the user running the node app)
to have root privileges, but we also don't want the /proc/cplayground
file to be readable by any user. I finally figured out how to change
ownership of the file so that the cplayground user can read it (and only
the cplayground user can read it).