Fix race condition causing containerId to be improperly set
This fixes an issue that has been around for a while (but which I just
noticed) where, for a fraction of a second, containerId was being
improperly set to a ChildProcess object before being resolved to the
correct string. I'd accidentally set containerId =
child_process.execFile(...) in addition to (properly) setting it in the
execFile callback.