Diagrams frontend: re-disable dragging of nodes
Dragging behavior is not currently well defined, and I don't know that
it's all that useful as of now. Also, it makes life hard with the way
we're currently generating the diagrams (we re-generate the whole
diagram every second, so if someone dragged, the node positions all get
reset less than a second later).