Store program hash in separate column; change id to auto-inc integer
I realized it's not a great idea to have program hashes as their primary
key IDs, because if the hash ever changes (because we add more things
that can be stored with a program, or change how options are stored) it
is a huge pain to regenerate IDs and fix all the foreign key references.
This commit rebuilds the ID column as an auto-incrementing integer
(might as well do it now while we're small), and stores the hashes in a
separate `hash` column.