Reimplement file uploads
As the last outstanding issue with this refactor, I've reimplemented
file uploads. Previously, file contents were being stored in a separate
column with each program. That means that if you modified a program
several times, the file would be uploaded and stored (in full)
separately for each program. Not great, but it was a quick fix. In this
commit, I've changed things so that files are stored in a separate
table, and programs store a reference to the file that they use. When
the user selects a file to include with their program, the JS will
immediately send the file to the server for upload. The server will
respond with a file ID. Then, when the user goes to run the program, the
JS will submit the program along with the file ID. If the user modifies
the program and reruns, the file is not re-uploaded; the ID is merely
re-submitted.

There are a few parts to this commit:
* A somewhat hairy database migration which makes this change to the
  database structure. Program hashes have to be regenerated, since we're
  changing the data that is stored with programs.
* On the frontend, when the user selects a file, a loading spinner is
  shown while the file is (immediately) uploaded. The returned file ID
  is stored in state and submitted along with the program when the run
  button is clicked.
* On the backend, added the endpoint to upload a file. Also made some
  changes to the database layer to retrieve the file names from the file
  table when programs are retrieved