[DB] Use hash as id for files table
Previously, I was using UUIDs as ids for the files table, and used a
UNIQUE constraint to guarantee that no duplicate files would be stored.
However, this isn't great for performance (unique indexes don't have
great performance for big columns like mediumblob), and, more
importantly, some MySQL versions don't support indexing on mediumblob at
all. This commit replaces the UUID ID with a SHA-2 224 hash of the
filename and contents.