Convert run.sh to python script
We're outgrowing the limits of the existing bash script just a bit. For
example, we want to grant ptrace capabilities to the container but
ensure that the user's program never gets them, and some of the code
seems better suited for a C/C++ launcher tbh. I figure a Python script
is a good in-between for now that will let us have a bit more
flexibility and control without having to rewrite in C/C++ or screw
around a lot with how bash does things.