Implement server API for debugger control
This commit adds server-side logic for setting/removing breakpoints and
controlling debugger execution (proceed, step in, step out, next). Two
fields are added to the Run event body, allowing the client to specify
whether the debugger should be used, as well as an initial set of
breakpoints. Also, this commit adds the '-g' compiler flag when
debugging is enabled.