Bugfix: Output not appearing, or appearing jumbled
This commit fixes a bug where the output was not appearing on some runs
(especially fast runs), and sometimes parts of output were appearing out
of order. The cause of this was that the container server code was
waiting for the container id query to finish before sending output to
the client. I'm not sure why I had that `await` (it shouldn't be
necessary, as the code is written to assume we don't necessarily know
the container ID), so I just got rid of it.