Client: Add ability to mark breakpoints (#13)
This commit adds the ability for users to set breakpoints in the client.
Clicking a line number will display a red circle next to that line, and
the line number will be added to a `breakpoints` array in the `App`
state. Clicking the line number a second time will remove the
breakpoint. Breakpoints stay anchored to lines of code, so they will
move around appropriately if you add/remove code earlier in the file.

![](https://user-images.githubusercontent.com/3021974/69368976-fd4f3e00-0c4f-11ea-9f7a-ada1aa0fe5ca.png)

Co-authored-by: Blanca Villanueva <blanca@cs.stanford.edu>
Co-authored-by: Ryan Eberhardt <reberhardt7@gmail.com>