Remove step out command
I'm going to remove the step out command, at least for now. It probably
won't get used that much, especially since we don't allow multiple
files, and it adds some complexity (you can't do "step out" in the main
function, which is where many people will have most of their code. We
have to figure out a way to disable the button in that context, which
isn't trivial).