Display debugger controls only for stopped processes
This commit modifies container.ts so that thread state is stored, and
communication.ts + debugging.ts so that this state is reported to the
client. ProcessesListing.tsx is modified so that the client only
displays the debugging controls if a process is stopped, and shows a
spinning loading icon otherwise (if the process is executing).