Send updated debugging info whenever a thread starts/stops
Before, we had a 1-second timer in debugging.ts that would poll
/proc/cplayground and parse/cache data, and we also had a 1-second timer
for each container that would pull from the cached procfile data and
send to the client. The purpose of this was to avoid reading the
procfile too often, since I don't have a good sense of how it performs
under load and it'll be bad if we overload it.

In this commit, I got rid of the debugging.ts poll mechanism and
replaced it with a better (but slightly complex) way to rate limit reads
to the procfile. I added a new class (Procfile) that keeps track of when
the last refresh was. If the last refresh was very recent, it adds the
refresh request to a queue. Eventually, when the procfile is next read,
all the promises in the queue are resolved. This way, the file is read
on demand, but won't be read more than once per
PROCFILE_MAX_REFRESH_PERIOD milliseconds.

Then, I added two lines to container.ts to request debugging data and
send it to the client every time an inferior starts or stops. This way,
the user will have a snappier experience when using the debugger
(instead of having up to 1s latency).