Client refactor: Receive debug data at App level
I'm working on adding debug controls inline in the Editor (so you can
easily see where threads are and control them easily), but that requires
that the Editor have access to debug data. As such, I'm having App
receive the data and pass it both to the Editor and to the Debugger
panel.

As part of this, I reorganized the socket management functions and debug
command functions under the DebugServer class. In the future, I would
like to do something similar for the rest of server-comm.ts, so that we
avoid passing the socket around as much as possible to ensure that
socket logic is confined to server-comm.ts.