Fix negative breakpoint bug
There was a bug where if you had a breakpoint set and then did cmd+a and
replaced all the code with new code, the existing breakpoints would
become negative line numbers. This would cause strange issues where the
editor would think that breakpoints were set, but gdb would reject the
breakpoints as invalid. This makes it impossible to set negative
breakpoints.