Switch from npm to yarn
I would prefer to keep using npm since it's the default and is built
into a lot of tooling, but it seems to have some bad interactions with
virtualbox shared folders. (I can't get "npm install" to work when
node_modules resides inside a Virtualbox shared directory, and after a
few hours of Googling, I haven't found a good solution.) Yarn works
fine, so I'll switch to it unless there is some reason not to.