Fix Vagrant build: only insmod if kernel mod exists
The Vagrant build was failing because `config.vm.provision :reload`
(which reboots the VM after installing the custom kernel but before
compiling the cplayground kernel module) was running all provision steps
marked `run: 'always'`, including the provisioning step to mount the
kernel module (which hasn't yet been compiled at this point). The simple
fix is to change that provisioning step to only mount the module if it
exists.