Show currently executing line for running threads
This commit modifies reportDebugInfo (called whenever a thread
stops/continues/etc, and also called every second or two, to send
debugging state to the client) so that it sends more state info about
running threads. It interrupts each running thread, gets a stack trace
to figure out which function the thread is running, then promptly
continues the thread, hopefully with no one noticing that anything even
happened. The client is updated to show the currently executing line in
the editor pane for all threads.