Debugger: show pending/blocked signals, allow sending signals
This commit adds a new Signals tab to the debugger. Pending/blocked
signal information is sourced via the kernel module. (I tried doing it
via gdb first, but it was too clunky... you can only get the info when
the process is paused, so in order to display it in real time for
running processes, you need to periodically pause/resume the child
processes, and it gets messy. Also, I couldn't even find a clean way to
get all pending signals... so this way was just simpler.

Also:
* Add a new debugSendSignal websocket event for sending a signal
* Add a defaultDebugTab GET param that specifies the default tab to
  open when debugging. This will be super non-obvious and there is
  nothing to suggest to users that this feature exists, but it will be
  really useful for me in lecture notes / embedded exercises. More UX
  to follow.