Fix GDB hanging on multithreaded programs
It seems that `advance`/`until` (used to advance to the next line, as a
replacement for the `next` command which doesn't work well when
interrupting/resuming from standard library functions) are the culprit
causing GDB to lock up in multithreaded programs. I'm switching back to
ordinary `next`, which doesn't work great when stepping line-by-line,
but it's better than freezing.